+++ connections.js
@@ -21,6 +21,8 @@ module.exports = [
   'data:',
   'https://app.lokalise.com:443',
   'https://myetherwallet.matomo.cloud:443',
+  'https://raw.githubusercontent.com:443/MyEtherWallet/dynamic-data/main/careers.json',
+  'https://formspree.io:443/f/mqjndkkx',
   // nodes
   'wss://nodes.mewapi.io:443',
   'wss://rpc.gochain.io:443/ws',