+++ src/app.vue
@@ -42,6 +42,8 @@ export default {
     });
   },
   mounted() {
+    this.footerHideIntercom();
+    this.logMessage();
     this.setOnlineStatus(window.navigator.onLine);
     if (window.navigator.onLine) {
       this.setCurrency(currencyTypes.USD);
@@ -64,7 +66,34 @@ export default {
   methods: {
     ...mapActions('global', ['setOnlineStatus']),
     ...mapActions('external', ['setCurrency']),
-    ...mapActions('addressBook', ['setMigrated', 'setAddressBook'])
+    ...mapActions('addressBook', ['setMigrated', 'setAddressBook']),
+    logMessage() {
+      /* eslint-disable no-console */
+      console.log(
+        '%cWhoa whoa whoa!',
+        'font-weight: bold',
+        '\n\nThis feature is intended only for developers.  Using it without knowing exactly what you are doing can expose your wallet keys and lead to the loss of your funds.',
+        '\n\nOn the other hand, if you are a developer and do know what you’re doing, MEW is hiring and we probably want to talk to you. Send us an email at careers@myetherwallet.com with the subject line: I am a software developer.'
+      );
+      /* eslint-enable no-console */
+    },
+    // Hide intercom button when users reach the footer or bottom of screen
+    footerHideIntercom() {
+      window.onscroll = function () {
+        if (
+          window.innerHeight + window.scrollY >=
+          document.body.offsetHeight - 100
+        ) {
+          window.Intercom('update', {
+            hide_default_launcher: true
+          });
+        } else {
+          window.Intercom('update', {
+            hide_default_launcher: false
+          });
+        }
+      };
+    }
   }
 };
 </script>