+++ src/dapps/ens-manager-dapp/TheENSManagerLayout.vue
@@ -19,10 +19,9 @@
     -->
       <template #tabContent1>
         <v-sheet
-          min-height="500px"
           max-width="700px"
           color="transparent"
-          class="py-15 mx-auto"
+          class="px-3 py-8 py-md-13 mx-auto"
         >
           <div class="mb-5">
             <div class="mew-heading-2 mb-8 ml-2">
@@ -68,7 +67,7 @@
     =====================================================================================
     -->
       <template #tabContent2>
-        <v-sheet min-height="500px" class="pa-3 pa-md-12">
+        <v-sheet class="px-3 py-8 py-md-13">
           <div class="d-flex align-center justify-space-between mb-7">
             <span class="mew-heading-2 font-weight-bold">
               {{ $t('ens.my-domains') }}
@@ -196,6 +195,44 @@
           </mew-expand-panel>
         </v-sheet>
       </template>
+      <template #tabContent3>
+        <v-sheet
+          max-width="500px"
+          color="transparent"
+          class="px-3 py-8 py-md-13 mx-auto"
+        >
+          <div class="mb-5">
+            <!--
+            ===================================================
+              Claim TITLE: hasEnsTokenBalance
+            ===================================================
+            -->
+            <claim-balance
+              :balance="ensTokens.balance"
+              :claimed="ensTokens.claimed"
+            />
+            <form
+              v-if="!ensTokens.claimed && hasEnsTokenBalance"
+              @submit.prevent="claimTokens"
+            >
+              <module-address-book
+                :label="$t('ens.delegator.addr')"
+                :is-valid-address-func="isValidDelegatorAddress"
+                preselect-curr-wallet-adr
+                @setAddress="setDelegatorAddress"
+              />
+              <mew-button
+                :loading="loading"
+                :disabled="isClaimDisabled"
+                :has-full-width="true"
+                btn-size="xlarge"
+                :title="$t('ens.claim-token-title')"
+                @click.native="claimTokens"
+              />
+            </form>
+          </div>
+        </v-sheet>
+      </template>
     </the-wrapper-dapp>
     <!--
     =====================================================================================
@@ -250,17 +287,27 @@ import ensBannerImg from '@/assets/images/backgrounds/bg-ens.png';
 import ModuleRegisterDomain from './modules/ModuleRegisterDomain';
 import ModuleManageDomain from './modules/ModuleManageDomain';
 import handlerEnsManager from './handlers/handlerEnsManager';
+import ClaimBalance from './components/claim/ClaimBalance';
 import { mapGetters, mapState } from 'vuex';
 import { Toast, ERROR, SUCCESS } from '@/modules/toast/handler/handlerToast';
 import BigNumber from 'bignumber.js';
 import ENS from 'ethereum-ens';
-import { fromWei } from 'web3-utils';
+import { fromWei, toBN } from 'web3-utils';
 import { formatIntegerToString } from '@/core/helpers/numberFormatHelper';
 import { ROUTES_WALLET } from '@/core/configs/configRoutes';
 import normalise from '@/core/helpers/normalise';
+import { isAddress } from '@/core/helpers/addressUtils';
+import ModuleAddressBook from '@/modules/address-book/ModuleAddressBook';
+import { hasClaimed, submitClaim } from './handlers/handlerENSTokenClaim';
 export default {
   name: 'ENSManagerLayout',
-  components: { ModuleRegisterDomain, ModuleManageDomain, TheWrapperDapp },
+  components: {
+    ModuleRegisterDomain,
+    ModuleManageDomain,
+    TheWrapperDapp,
+    ModuleAddressBook,
+    ClaimBalance
+  },
   data() {
     return {
       activeTab: 0,
@@ -276,6 +323,11 @@ export default {
       ensManager: {},
       onRegister: false,
       searchError: '',
+      ensTokens: {
+        claimed: false,
+        balance: '0',
+        proof: ''
+      },
       manageDomainOptions: [
         {
           label: this.$t('ens.transfer-domain'),
@@ -305,23 +357,44 @@ export default {
       ],
       tabs: [
         { name: this.$t('ens.register-domain') },
-        { name: this.$t('ens.manage-domain') }
+        { name: this.$t('ens.manage-domain') },
+        { name: this.$t('ens.claim-tokens') }
       ],
       myDomains: [],
       ensBannerImg: ensBannerImg,
       bannerText: {
         title: this.$t('ens.title'),
         subtext: this.$t('ens.dapp-desc')
-      }
+      },
+      delegatorAddress: ''
     };
   },
   computed: {
     ...mapGetters('global', ['network', 'gasPrice']),
     ...mapGetters('external', ['fiatValue']),
-    ...mapState('wallet', ['balance', 'address', 'web3']),
+    ...mapState('wallet', ['balance', 'address', 'web3', 'instance']),
+    hasEnsTokenBalance() {
+      if (this.ensTokens.balance) {
+        return toBN(this.ensTokens.balance).gt(toBN(0));
+      }
+      return false;
+    },
     errorMessages() {
       if (this.domainTaken) return this.$t('ens.domain-taken');
       return this.searchError;
+    },
+    delegatorErrors() {
+      if (!isAddress(this.delegatorAddress)) {
+        return 'Invalid address!';
+      }
+      return '';
+    },
+    isClaimDisabled() {
+      return (
+        this.ensTokens.claimed ||
+        this.delegatorErrors !== '' ||
+        this.delegatorAddress === ''
+      );
     },
     rules() {
       return [
@@ -369,6 +442,19 @@ export default {
       if (newVal === true) {
         this.onRegister = true;
       }
+    },
+    /* 
+    - watches for address state change
+    - updates ensManager with new address 
+    - if user is onRegister it will reset and take them back
+    - if user is onManage it will run getDomain to refresh domains
+    */
+    address(newVal) {
+      this.ensManager.address = newVal;
+      if (this.onRegister) {
+        this.closeRegister();
+      }
+      this.getDomains();
     }
   },
   mounted() {
@@ -384,8 +470,31 @@ export default {
     );
 
     this.getDomains();
+    hasClaimed(this.address, this.web3).then(data => {
+      this.ensTokens.claimed = data.claimed;
+      this.ensTokens.balance = data.balance;
+      this.ensTokens.proof = data.proof;
+    });
   },
   methods: {
+    claimTokens() {
+      try {
+        submitClaim(
+          this.ensTokens.balance,
+          this.ensTokens.proof,
+          this.delegatorAddress,
+          this.web3
+        ).catch(this.instance.errorHandler);
+      } catch (e) {
+        this.instance.errorHandler(e);
+      }
+    },
+    setDelegatorAddress(address) {
+      this.delegatorAddress = address;
+    },
+    isValidDelegatorAddress(address) {
+      return isAddress(address);
+    },
     buyDomain() {
       this.activeTab = 0;
     },
@@ -431,7 +540,7 @@ export default {
           }, 15000);
         })
         .catch(err => {
-          Toast(err, {}, ERROR);
+          this.instance.errorHandler(err);
         });
       this.closeManage();
     },
@@ -440,7 +549,7 @@ export default {
         .renew(duration, this.balanceToWei)
         .then(this.getDomains)
         .catch(err => {
-          Toast(err, {}, ERROR);
+          this.instance.errorHandler(err);
         });
       this.closeManage();
     },
@@ -449,7 +558,7 @@ export default {
         .setMulticoin(coin)
         .then(this.getDomains)
         .catch(err => {
-          Toast(err, {}, ERROR);
+          this.instance.errorHandler(err);
         });
       this.closeManage();
     },
@@ -458,7 +567,7 @@ export default {
         .setTxtRecord(records)
         .then(this.getDomains)
         .catch(err => {
-          Toast(err, {}, ERROR);
+          this.instance.errorHandler(err);
         });
       this.closeManage();
     },
@@ -475,7 +584,7 @@ export default {
           this.closeManage();
         })
         .catch(err => {
-          Toast(err, {}, ERROR);
+          this.instance.errorHandler(err);
         });
     },
     setIpfs(hash) {
@@ -486,7 +595,7 @@ export default {
           this.settingIpfs = false;
         })
         .catch(err => {
-          Toast(err, {}, ERROR);
+          this.instance.errorHandler(err);
         });
       this.closeManage();
     },
@@ -535,7 +644,7 @@ export default {
           }, 15000);
         })
         .on('error', err => {
-          Toast(err, {}, ERROR);
+          this.instance.errorHandler(err);
         });
     },
     commit() {
@@ -557,7 +666,6 @@ export default {
           }, waitingTime * 1000);
         })
         .on('error', err => {
-          this.closeRegister();
           Toast(err, {}, ERROR);
         });
     },
@@ -594,5 +702,9 @@ export default {
       }
     }
   }
+}
+
+.claim-border-container {
+  border: 1px solid var(--v-greyMedium-base);
 }
 </style>