+++ src/modules/access-wallet/ModuleAccessWalletHardware.vue
@@ -71,7 +71,7 @@
               Step 2: Start Access to Selected Hardware Wallet
             =====================================================================================
             -->
-    <div v-if="step === 2" class="full-width">
+    <div v-if="step <= walletInitialized" class="full-width">
       <!--
         =====================================================================================
           Bitbox2
@@ -126,6 +126,7 @@
         :paths="paths"
         :selected-path="selectedPath"
         :set-path="setPath"
+        @ledgerApp="setSelectedApp"
       />
 
       <!--
@@ -139,7 +140,7 @@
     </div>
     <!--
       =====================================================================================
-        Step 3: Select Address and Network | (If Applicable) 
+        Step 3: Select Address and Network | (If Applicable)
       =====================================================================================
       -->
     <!--
@@ -148,7 +149,7 @@
     =====================================================================================
     -->
     <access-wallet-address-network
-      v-if="step === 3"
+      v-if="step > walletInitialized"
       :back="null"
       :hide-custom-paths="onKeepkey || onLedger"
       :handler-wallet="hwWalletInstance"
@@ -163,7 +164,7 @@
 
 <script>
 import { Toast, ERROR } from '@/modules/toast/handler/handlerToast';
-import { isEmpty } from 'lodash';
+import { isEmpty, isObject } from 'lodash';
 import AccessWalletBitbox from './hardware/components/AccessWalletBitbox';
 import AccessWalletAddressNetwork from '@/modules/access-wallet/common/components/AccessWalletAddressNetwork';
 import AccessWalletKeepkey from './hardware/components/AccessWalletKeepkey';
@@ -270,6 +271,11 @@ export default {
   computed: {
     ...mapGetters('global', ['Networks', 'network']),
     ...mapState('wallet', ['identifier']),
+    walletInitialized() {
+      return this.wallets[this.walletType]
+        ? this.wallets[this.walletType]?.when
+        : 1;
+    },
     /**
      * Returns the correct network icon
      */
@@ -411,7 +417,7 @@ export default {
      */
     title() {
       if (this.switchAddress) return 'Switch Address';
-      if (this.step > this.wallets[this.walletType]?.when) {
+      if (this.step > this.walletInitalzed) {
         return 'Select Network and Address';
       } else if (this.step === 1) {
         return 'Select a hardware wallet';
@@ -452,10 +458,10 @@ export default {
         : false;
     },
     bitbox2Titles() {
-      if (this.bitBox2Connected) return 'Enter Bitbox 02 password';
+      if (this.bitBox2Connected) return 'Enter BitBox02 password';
       if (this.bitBox2Unpaired) return 'Confirm pairing code';
       if (this.bitBox2Initialized)
-        return 'Bitbox 02 succesfully initialized. Loading wallet';
+        return 'BitBox02 succesfully initialized. Loading wallet';
       return this.walletType ? this.wallets[this.walletType].title : '';
     }
   },
@@ -514,7 +520,7 @@ export default {
         if (this.step === 1) {
           this.reset();
         } else if (this.step === 2) {
-          this.step -= 1;
+          this.step = 1;
         } else {
           this.hwWalletInstance = {};
           if (this.onLedger) {
@@ -544,7 +550,7 @@ export default {
     nextStep() {
       if (this.walletType) {
         this.step++;
-        if (this.step === this.wallets[this.walletType].when) {
+        if (this.step === this.walletInitialized) {
           if (this.onLedger) this.selectedPath = this.paths[0];
           if (this.onCoolWallet || this.onBitbox2) return;
           this[`${this.walletType}Unlock`]();
@@ -573,9 +579,11 @@ export default {
      * Unlock only the path step
      */
     unlockPathOnly() {
-      const path = this.selectedPath.hasOwnProperty('value')
-        ? this.selectedPath.value
-        : this.selectedPath;
+      const path = isObject(this.selectedPath)
+        ? this.selectedPath.hasOwnProperty('value')
+          ? this.selectedPath.value
+          : this.selectedPath
+        : this.paths[0];
       this.wallets[this.walletType]
         .create(path)
         .then(_hwWallet => {
@@ -586,6 +594,12 @@ export default {
               this.step++;
             if (this.onBitbox2) {
               this.hwWalletInstance = _hwWallet;
+              if (!this.hwWalletInstance) {
+                this.wallets[this.walletType].create.errorHandler(
+                  'bitboxInstanceError'
+                );
+                return;
+              }
               this.hwWalletInstance
                 .init()
                 .then(() => {
@@ -678,6 +692,12 @@ export default {
     setPassword(str) {
       this.password = str;
       this.passwordError = false;
+    },
+    /**
+     * Sets selected app for Ledger
+     */
+    setSelectedApp(e) {
+      this.selectedLedgerApp = e;
     }
   }
 };