+++ src/modules/access-wallet/hardware/components/AccessWalletBitbox.vue
@@ -6,7 +6,7 @@
       </v-col>
       <v-col cols="12" justify="center" align="center">
         <mew-button
-          title="Unlock Bitbox 02"
+          title="Unlock BitBox02"
           color-theme="primary"
           :has-full-width="true"
           btn-size="xlarge"