+++ src/modules/confirmation/components/ConfirmationValuesContainer.vue
@@ -50,7 +50,9 @@
           <div class="mew-caption textLight--text pb-2">
             {{ item.title }}
           </div>
-          <mew-token-container :icon="item.icon" size="medium" />
+          <div class="d-flex align-center justify-center">
+            <mew-token-container :img="item.icon" size="medium" />
+          </div>
           <div class="mew-heading-4 font-weight-medium black--text pt-2">
             {{ item.value }}
             <span class="mew-body font-weight-medium"> {{ item.type }} </span>