+++ src/modules/settings/components/SettingsGasPrice.vue
@@ -114,12 +114,13 @@
 </template>
 
 <script>
-import { toBN, fromWei } from 'web3-utils';
+import { fromWei } from 'web3-utils';
 import { gasPriceTypes } from '@/core/helpers/gasPriceHelper';
 import { mapState, mapGetters } from 'vuex';
 import {
   formatFiatValue,
-  formatFloatingPointValue
+  formatFloatingPointValue,
+  toBNSafe
 } from '@/core/helpers/numberFormatHelper';
 import BigNumber from 'bignumber.js';
 
@@ -243,8 +244,10 @@ export default {
   methods: {
     calcTxFee(priority) {
       return fromWei(
-        toBN(this.totalGasLimit).mul(toBN(this.gasPriceByType(priority)))
-      ).toString();
+        toBNSafe(this.totalGasLimit)
+          .mul(toBNSafe(this.gasPriceByType(priority)))
+          .toString()
+      );
     },
     formatInEth(fee) {
       return formatFloatingPointValue(fee).value;