+++ src/modules/swap/ModuleSwap.vue
@@ -14,7 +14,7 @@
             =====================================================================================
             -->
           <div class="input-swap-container pt-7 pb-3 px-5">
-            <v-row class="align-center justify-space-between mt-4">
+            <v-row class="align-start justify-space-between mt-4">
               <v-col cols="12" sm="5" class="pb-0 pb-sm-3 pr-sm-0">
                 <div class="position--relative">
                   <app-button-balance
@@ -53,20 +53,24 @@
                   }"
                   @input="setTokenInValue"
               /></v-col>
-              <v-col cols="12" sm="2" class="px-6 py-0 py-sm-3 mb-3 mb-sm-0">
+              <v-col
+                cols="12"
+                align-self="center"
+                sm="2"
+                class="px-6 py-0 py-sm-3 mb-3 mb-sm-0"
+              >
                 <div class="d-flex align-center justify-center pb-sm-10">
-                  <swap-btn
-                    :class="[
-                      enableTokenSwitch
-                        ? 'cursor--pointer'
-                        : 'pointer-event--none',
-                      'd-flex align-center justify-center'
-                    ]"
+                  <mew-icon-button
+                    mdi-icon="swap-horizontal"
+                    class="pa-2 d-flex align-center justify-center"
+                    color-theme="basic"
+                    btn-style="light"
+                    :disabled="!enableTokenSwitch"
                     @click.native="switchTokens"
                   />
                 </div>
               </v-col>
-              <v-col cols="12" sm="5" class="pl-sm-0 pb-0 pb-sm-3">
+              <v-col cols="12" sm="5" class="pb-0 pb-sm-3 pl-sm-0">
                 <mew-select
                   ref="toToken"
                   :value="toTokenType"
@@ -281,7 +285,6 @@
 </template>
 
 <script>
-import SwapBtn from '@/views/components-wallet/TheSwapBtn';
 import AppButtonBalance from '@/core/components/AppButtonBalance';
 import AppUserMsgBlock from '@/core/components/AppUserMsgBlock';
 import ModuleAddressBook from '@/modules/address-book/ModuleAddressBook';
@@ -308,7 +311,6 @@ const MIN_GAS_LIMIT = 800000;
 export default {
   name: 'ModuleSwap',
   components: {
-    SwapBtn,
     AppButtonBalance,
     AppUserMsgBlock,
     ModuleAddressBook,
@@ -470,10 +472,10 @@ export default {
      */
     enableTokenSwitch() {
       return (
-        !isEmpty(this.fromTokenType) &&
-        !isEmpty(this.toTokenType) &&
-        !isEmpty(this.fromTokenType?.symbol) &&
-        !isEmpty(this.toTokenType?.symbol)
+        !this.isLoading &&
+        ((!isEmpty(this.fromTokenType) &&
+          !isEmpty(this.fromTokenType?.symbol)) ||
+          (!isEmpty(this.toTokenType) && !isEmpty(this.toTokenType?.symbol)))
       );
     },
     /**
@@ -664,7 +666,11 @@ export default {
       return this.addressValue.value;
     },
     isToAddressValid() {
-      if (this.toTokenType?.isEth) return true;
+      if (
+        this.toTokenType?.isEth ||
+        this.toTokenType?.contract === MAIN_TOKEN_ADDRESS
+      )
+        return true;
       return this.addressValue.isValid;
     },
     /**
@@ -815,16 +821,16 @@ export default {
       deep: true,
       immediate: true
     },
-    network() {
-      if (this.isAvailable) {
-        this.clear();
-      }
-    },
     mainTokenDetails() {
       this.setDefaults();
     },
     amountErrorMessage(newVal) {
       if (newVal !== '') this.availableQuotes.splice(0);
+    },
+    '$route.query': {
+      handler: function () {
+        this.setTokenFromURL();
+      }
     }
   },
   beforeMount() {
@@ -890,8 +896,8 @@ export default {
       this.addressValue = {};
       this.selectedProvider = {};
       this.localGasPrice = '0';
-      this.$refs.toToken.clear();
-      this.$refs.amountInput.clear();
+      if (this.$refs.toToken) this.$refs.toToken.clear();
+      if (this.$refs.amountInput) this.$refs.amountInput.clear();
       this.setupSwap();
     },
     formatTokensForSelect(tokens) {
@@ -961,6 +967,7 @@ export default {
     switchTokens() {
       const fromToken = clone(this.fromTokenType);
       const toToken = clone(this.toTokenType);
+      this.tokenInValue = '0';
       this.setFromToken(toToken);
       this.setToToken(fromToken);
     },
@@ -1044,7 +1051,10 @@ export default {
       if (
         this.tokenInValue !== '' &&
         this.tokenInValue > 0 &&
-        this.enableTokenSwitch
+        !isEmpty(this.fromTokenType) &&
+        !isEmpty(this.toTokenType) &&
+        !isEmpty(this.fromTokenType?.symbol) &&
+        !isEmpty(this.toTokenType?.symbol)
       ) {
         this.isLoadingProviders = true;
         this.showAnimation = true;
@@ -1092,7 +1102,8 @@ export default {
       this.step = 1;
       this.feeError = '';
       this.loadingFee = true;
-      if (this.allTrades[idx]) return this.setupTrade(this.allTrades[idx]);
+      if (this.allTrades.length > 0 && this.allTrades[idx])
+        return this.setupTrade(this.allTrades[idx]);
       const trade = this.swapper.getTrade({
         fromAddress: this.address,
         toAddress: this.toAddress,
@@ -1105,26 +1116,23 @@ export default {
         )
       });
       if (trade instanceof Promise) {
-        trade
-          .then(tradeResponse => {
-            this.allTrades[idx] = tradeResponse;
-            this.setupTrade(tradeResponse);
-          })
-          .catch(e => {
-            if (e) {
-              this.feeError = 'This provider is not available.';
-            }
-          });
+        trade.then(tradeResponse => {
+          this.allTrades[idx] = tradeResponse;
+          this.setupTrade(tradeResponse);
+        });
+      } else {
+        this.setupTrade(trade);
       }
     }, 500),
     setupTrade(trade) {
-      if (trade instanceof Error) {
+      this.step = 2;
+      if (trade instanceof Error || !trade) {
         this.feeError = 'Provider issue';
+        this.loadingFee = false;
         return;
       }
       this.currentTrade = trade;
       this.currentTrade.gasPrice = this.currentGasPrice;
-      this.step = 2;
       this.loadingFee = false;
       this.checkFeeBalance();
     },
@@ -1325,5 +1333,13 @@ export default {
 .swap-to-input {
   pointer-events: none !important;
   user-select: none !important;
+}
+
+.pb-sm-15px {
+  padding-bottom: 15px !important;
+}
+
+.pb-sm-29px {
+  padding-bottom: 29px !important;
 }
 </style>