+++ src/modules/swap/ModuleSwapRates.vue
@@ -11,7 +11,7 @@
         />
       </div>
     </div>
-    <div v-if="!loading && !error" class="pa-3">
+    <div v-if="!loading && !error && hasSwapRates" class="pa-3">
       <div v-for="(data, key) in swapData" :key="key">
         <v-sheet
           v-if="data.rate"
@@ -59,12 +59,14 @@
       <h3 class="ma-3">Loading swap pairs...</h3>
     </div>
     <div
-      v-if="error"
+      v-if="error || !hasSwapRates"
       class="pa-3 pb-4 d-flex flex-column align-center justify-space-around"
     >
       <v-progress-circular indeterminate />
       <h3 class="ma-3">Having issues loading tokens.</h3>
-      <h5 class="mb-2 cursor--pointer" @click="fetchRates">Try again?</h5>
+      <h5 class="mb-2 cursor--pointe greenPrimary--text" @click="fetchRates">
+        Try again?
+      </h5>
     </div>
   </mew6-white-sheet>
 </template>
@@ -75,6 +77,7 @@ import { mapState, mapGetters } from 'vuex';
 import { formatFloatingPointValue } from '@/core/helpers/numberFormatHelper';
 import { Toast, ERROR } from '@/modules/toast/handler/handlerToast';
 import handlerAnalytics from '@/modules/analytics-opt-in/handlers/handlerAnalytics.mixin';
+import { toWei } from 'web3-utils';
 
 const STATIC_PAIRS = [
   {
@@ -87,7 +90,7 @@ const STATIC_PAIRS = [
       contract: '0xeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeee',
       decimals: 18
     },
-    fromAmount: '100000000000000000'
+    fromAmount: toWei('1')
   },
   {
     fromT: {
@@ -100,7 +103,7 @@ const STATIC_PAIRS = [
       contract: '0xdac17f958d2ee523a2206206994597c13d831ec7',
       decimals: 6
     },
-    fromAmount: '1000000000000000000'
+    fromAmount: toWei('1')
   },
   {
     fromT: {
@@ -113,7 +116,7 @@ const STATIC_PAIRS = [
       contract: '0xdd974d5c2e2928dea5f71b9825b8b646686bd200',
       toT: 18
     },
-    fromAmount: '100000000000000000'
+    fromAmount: toWei('1')
   },
   {
     fromT: {
@@ -126,7 +129,7 @@ const STATIC_PAIRS = [
       contract: '0x6b175474e89094c44da98b954eedeac495271d0f',
       decimals: 18
     },
-    fromAmount: '100000000000000000'
+    fromAmount: toWei('1')
   },
   {
     fromT: {
@@ -139,7 +142,7 @@ const STATIC_PAIRS = [
       contract: '0x514910771af9ca656af840dff83e8264ecf986ca',
       decimals: 18
     },
-    fromAmount: '100000000000000000'
+    fromAmount: toWei('1')
   },
   {
     fromT: {
@@ -152,10 +155,11 @@ const STATIC_PAIRS = [
       contract: '0xa0b86991c6218b36c1d19d4a2e9eb0ce3606eb48',
       decimals: 6
     },
-    fromAmount: '100000000000000000'
+    fromAmount: toWei('1')
   }
 ];
 export default {
+  name: 'ModuleSwapRates',
   components: {},
   mixins: [handlerAnalytics],
   props: {
@@ -174,15 +178,23 @@ export default {
   },
   computed: {
     ...mapState('wallet', ['web3']),
-    ...mapGetters('global', ['isEthNetwork'])
+    ...mapGetters('global', ['isEthNetwork', 'network']),
+    hasSwapRates() {
+      if (this.swapData) {
+        return this.swapData.some(item => {
+          return item.rate;
+        });
+      }
+      return false;
+    }
   },
   watch: {
     web3(newVal) {
-      this.setSwapHandler(newVal);
+      this.setSwapHandler(newVal, this.network.type.name);
     }
   },
   mounted() {
-    this.setSwapHandler(this.web3);
+    this.setSwapHandler(this.web3, this.network.type.name);
   },
   methods: {
     setSwapHandler(val) {